various changes, attempting to improve portability; still, there remain lots of open issues (see test-web for details).
Details:

- conf/conf.bash:
   * clean-up

- configure.ag:
   * check the availability of additional header files (stdbool.h, libgen.h),
     functions (basename), and types (bool, socklen_t)

- compiler/main.c:
   * ensure that <pf-config.h> is the first include
   * include <libgen.h> only if available
   * use basename() only if available
   * comment-out unused variable "pathname"

- compiler/daemon/daemon.c:
   * use "size_t" if "socklen_t" is not available
   * fix definition of "SIG_IGN" for picky Intel compiler on Linux64

- compiler/debug/prettyp.c:
   * "isspace()" expects an "int", not a "char"
   * include <stdbool.h> only if available,
   * otherwise, define "bool" if available

- compiler/mem/Makefile.ag:
   * add $(LIBGC_CFLAGS) to INCLUDES to ensure that gc.h is found

- compiler/mem/array.c:
   * make macro ACHUNK an unsigned int
   * avoid void-pointer arithmetic; some compilers don't like it

- compiler/mil/core2mil.c:
   * comment-out unused variable "w"

- loader/pf-shred.c:
   * include <libgen.h> only if available

- runtime/levelsteps.c:
   * comment-out unused variable "level_last"

- runtime/staircasejoin.mx:
   * extended casting to pacify 64-bit compilers

+++ pathfinder/compiler/debug/prettyp.c
@@ -276,8 +276,8 @@ receive (void)
     case END_BLOCK:
         return &end_block;
     default:
-        if (isspace (*s)) {
-            while (isspace (*((char *) PFarray_add (collect))))
+        if (isspace ((int)(*s))) {
+            while (isspace ((int)(*((char *) PFarray_add (collect)))))
                 ;
             PFarray_last (collect)--;
             return &blank;
@@ -288,7 +288,7 @@ receive (void)
             /* a string, s marks it start, e marks its end */
             do 
                 e = (char *) PFarray_add (collect);
-            while (*e && ! isspace (*e) && *e != START_BLOCK && *e != END_BLOCK);
+            while (*e && ! isspace ((int)(*e)) && *e != START_BLOCK && *e != END_BLOCK);
             PFarray_last (collect)--;
 
             /* copy the string and return it */
